Moved home page data to separate file
